Added documentation for {{partial}}.
